+++ packages/graphics/src/GraphicsGeometry.js
@@ -1,11 +1,8 @@
-import { Buffer, Geometry, Texture } from '@pixi/core';
-import { Circle, Ellipse, PI_2, Polygon, Rectangle, RoundedRectangle, SHAPES } from '@pixi/math';
+import { Buffer, Geometry } from '@pixi/core';
+import { Rectangle, SHAPES } from '@pixi/math';
 import { TYPES } from '@pixi/constants';
 
-import FillStyle from './styles/FillStyle';
 import GraphicsData from './GraphicsData';
-import LineStyle from './styles/LineStyle';
-import bezierCurveTo from './utils/bezierCurveTo';
 import buildCircle from './utils/buildCircle';
 import buildLine from './utils/buildLine';
 import buildPoly from './utils/buildPoly';
@@ -67,11 +64,6 @@ export default class GraphicsGeometry extends Geometry
             // then add the extra attribute!
         }
 
-        this._fillStyle = null;
-        this._lineStyle = null;
-
-        this.matrix = null;
-
         /**
          * Graphics data
          *
@@ -80,14 +72,6 @@ export default class GraphicsGeometry extends Geometry
          */
         this.graphicsData = [];
         this.graphicsDataHoles = [];
-
-        /**
-         * Current path
-         *
-         * @member {PIXI.GraphicsData}
-         * @private
-         */
-        this.currentPath = null;
 
         /**
          * Used to detect if the graphics object has changed. If this is set to true then the graphics
@@ -138,562 +122,31 @@ export default class GraphicsGeometry extends Geometry
     }
 
     /**
-     * Specifies the line style used for subsequent calls to Graphics methods such as the lineTo()
-     * method or the drawCircle() method.
-     *
-     * @param {number} [lineWidth=0] - width of the line to draw, will update the objects stored style
-     * @param {number} [color=0] - color of the line to draw, will update the objects stored style
-     * @param {number} [alpha=1] - alpha of the line to draw, will update the objects stored style
-     * @param {number} [alignment=1] - alignment of the line to draw, (0 = inner, 0.5 = middle, 1 = outter)
-     * @return {PIXI.Graphics} This Graphics object. Good for chaining method calls
-     */
-    lineStyle(width = 0, color = 0, alpha = 1, alignment = 0.5, native = false)
-    {
-        this.lineTextureStyle(width, Texture.WHITE, color, alpha, null, alignment, native);
-    }
-
-    lineTextureStyle(width = 0, texture, color = 0xFFFFFF, alpha = 1, textureMatrix, alignment = 0.5, native)
-    {
-        if (width === 0 || alpha === 0)
-        {
-            this._lineStyle = null;
-
-            return this;
-        }
-
-        const style = new LineStyle();
-
-        style.color = color;
-        style.width = width;
-        style.alpha = alpha;
-        style.matrix = textureMatrix;
-        style.texture = texture || Texture.WHITE;
-        style.alignment = alignment;
-        style.native = native;
-
-        if (this.currentPath)
-        {
-            if (this.currentPath.shape.points.length)
-            {
-                // halfway through a line? start a new one!
-                const shape = new Polygon(this.currentPath.shape.points.slice(-2));
-
-                shape.closed = false;
-
-                this.drawShape(shape, this._fillStyle, this._lineStyle, this.matrix);
-            }
-            else
-            {
-                // otherwise its empty so lets just set the line properties
-                this.currentPath.lineStyle = style;
-            }
-        }
-
-        this._lineStyle = style;
-
-        return this;
-    }
-
-    setMatrix(matrix)
-    {
-        this.matrix = matrix;
-    }
-
-    /**
-     * Moves the current drawing position to x, y.
-     *
-     * @param {number} x - the X coordinate to move to
-     * @param {number} y - the Y coordinate to move to
-     * @return {PIXI.Graphics} This Graphics object. Good for chaining method calls
-     */
-    moveTo(x, y)
-    {
-        const shape = new Polygon([x, y]);
-
-        shape.closed = false;
-        this.drawShape(shape, this._fillStyle, this._lineStyle, this.matrix);
-
-        return this;
-    }
-
-    /**
-     * Draws a line using the current line style from the current drawing position to (x, y);
-     * The current drawing position is then set to (x, y).
-     *
-     * @param {number} x - the X coordinate to draw to
-     * @param {number} y - the Y coordinate to draw to
-     * @return {PIXI.Graphics} This Graphics object. Good for chaining method calls
-     */
-    lineTo(x, y)
-    {
-        this.currentPath.shape.points.push(x, y);
-        this.dirty++;
-
-        return this;
-    }
-
-    /**
-     * Calculate the points for a quadratic bezier curve and then draws it.
-     * Based on: https://stackoverflow.com/questions/785097/how-do-i-implement-a-bezier-curve-in-c
-     *
-     * @param {number} cpX - Control point x
-     * @param {number} cpY - Control point y
-     * @param {number} toX - Destination point x
-     * @param {number} toY - Destination point y
-     * @return {PIXI.Graphics} This Graphics object. Good for chaining method calls
-     */
-    quadraticCurveTo(cpX, cpY, toX, toY)
-    {
-        if (this.currentPath)
-        {
-            if (this.currentPath.shape.points.length === 0)
-            {
-                this.currentPath.shape.points = [0, 0];
-            }
-        }
-        else
-        {
-            this.moveTo(0, 0);
-        }
-
-        const n = 20;
-        const points = this.currentPath.shape.points;
-        let xa = 0;
-        let ya = 0;
-
-        if (points.length === 0)
-        {
-            this.moveTo(0, 0);
-        }
-
-        const fromX = points[points.length - 2];
-        const fromY = points[points.length - 1];
-
-        for (let i = 1; i <= n; ++i)
-        {
-            const j = i / n;
-
-            xa = fromX + ((cpX - fromX) * j);
-            ya = fromY + ((cpY - fromY) * j);
-
-            points.push(xa + (((cpX + ((toX - cpX) * j)) - xa) * j),
-                ya + (((cpY + ((toY - cpY) * j)) - ya) * j));
-        }
-
-        this.dirty++;
-
-        return this;
-    }
-
-    /**
-     * Calculate the points for a bezier curve and then draws it.
-     *
-     * @param {number} cpX - Control point x
-     * @param {number} cpY - Control point y
-     * @param {number} cpX2 - Second Control point x
-     * @param {number} cpY2 - Second Control point y
-     * @param {number} toX - Destination point x
-     * @param {number} toY - Destination point y
-     * @return {PIXI.Graphics} This Graphics object. Good for chaining method calls
-     */
-    bezierCurveTo(cpX, cpY, cpX2, cpY2, toX, toY)
-    {
-        if (this.currentPath)
-        {
-            if (this.currentPath.shape.points.length === 0)
-            {
-                this.currentPath.shape.points = [0, 0];
-            }
-        }
-        else
-        {
-            this.moveTo(0, 0);
-        }
-
-        const points = this.currentPath.shape.points;
-
-        const fromX = points[points.length - 2];
-        const fromY = points[points.length - 1];
-
-        points.length -= 2;
-
-        bezierCurveTo(fromX, fromY, cpX, cpY, cpX2, cpY2, toX, toY, points);
-
-        this.dirty++;
-
-        return this;
-    }
-
-    /**
-     * The arcTo() method creates an arc/curve between two tangents on the canvas.
-     *
-     * "borrowed" from https://code.google.com/p/fxcanvas/ - thanks google!
-     *
-     * @param {number} x1 - The x-coordinate of the beginning of the arc
-     * @param {number} y1 - The y-coordinate of the beginning of the arc
-     * @param {number} x2 - The x-coordinate of the end of the arc
-     * @param {number} y2 - The y-coordinate of the end of the arc
-     * @param {number} radius - The radius of the arc
-     * @return {PIXI.Graphics} This Graphics object. Good for chaining method calls
-     */
-    arcTo(x1, y1, x2, y2, radius)
-    {
-        if (this.currentPath)
-        {
-            if (this.currentPath.shape.points.length === 0)
-            {
-                this.currentPath.shape.points.push(x1, y1);
-            }
-        }
-        else
-        {
-            this.moveTo(x1, y1);
-        }
-
-        const points = this.currentPath.shape.points;
-        const fromX = points[points.length - 2];
-        const fromY = points[points.length - 1];
-        const a1 = fromY - y1;
-        const b1 = fromX - x1;
-        const a2 = y2 - y1;
-        const b2 = x2 - x1;
-        const mm = Math.abs((a1 * b2) - (b1 * a2));
-
-        if (mm < 1.0e-8 || radius === 0)
-        {
-            if (points[points.length - 2] !== x1 || points[points.length - 1] !== y1)
-            {
-                points.push(x1, y1);
-            }
-        }
-        else
-        {
-            const dd = (a1 * a1) + (b1 * b1);
-            const cc = (a2 * a2) + (b2 * b2);
-            const tt = (a1 * a2) + (b1 * b2);
-            const k1 = radius * Math.sqrt(dd) / mm;
-            const k2 = radius * Math.sqrt(cc) / mm;
-            const j1 = k1 * tt / dd;
-            const j2 = k2 * tt / cc;
-            const cx = (k1 * b2) + (k2 * b1);
-            const cy = (k1 * a2) + (k2 * a1);
-            const px = b1 * (k2 + j1);
-            const py = a1 * (k2 + j1);
-            const qx = b2 * (k1 + j2);
-            const qy = a2 * (k1 + j2);
-            const startAngle = Math.atan2(py - cy, px - cx);
-            const endAngle = Math.atan2(qy - cy, qx - cx);
-
-            this.arc(cx + x1, cy + y1, radius, startAngle, endAngle, b1 * a2 > b2 * a1);
-        }
-
-        this.dirty++;
-
-        return this;
-    }
-
-    /**
-     * The arc method creates an arc/curve (used to create circles, or parts of circles).
-     *
-     * @param {number} cx - The x-coordinate of the center of the circle
-     * @param {number} cy - The y-coordinate of the center of the circle
-     * @param {number} radius - The radius of the circle
-     * @param {number} startAngle - The starting angle, in radians (0 is at the 3 o'clock position
-     *  of the arc's circle)
-     * @param {number} endAngle - The ending angle, in radians
-     * @param {boolean} [anticlockwise=false] - Specifies whether the drawing should be
-     *  counter-clockwise or clockwise. False is default, and indicates clockwise, while true
-     *  indicates counter-clockwise.
-     * @return {PIXI.Graphics} This Graphics object. Good for chaining method calls
-     */
-    arc(cx, cy, radius, startAngle, endAngle, anticlockwise = false)
-    {
-        if (startAngle === endAngle)
-        {
-            return this;
-        }
-
-        if (!anticlockwise && endAngle <= startAngle)
-        {
-            endAngle += PI_2;
-        }
-        else if (anticlockwise && startAngle <= endAngle)
-        {
-            startAngle += PI_2;
-        }
-
-        const sweep = endAngle - startAngle;
-        const segs = Math.ceil(Math.abs(sweep) / PI_2) * 40;
-
-        if (sweep === 0)
-        {
-            return this;
-        }
-
-        const startX = cx + (Math.cos(startAngle) * radius);
-        const startY = cy + (Math.sin(startAngle) * radius);
-
-        // If the currentPath exists, take its points. Otherwise call `moveTo` to start a path.
-        let points = this.currentPath ? this.currentPath.shape.points : null;
-
-        if (points)
-        {
-            if (points[points.length - 2] !== startX || points[points.length - 1] !== startY)
-            {
-                points.push(startX, startY);
-            }
-        }
-        else
-        {
-            this.moveTo(startX, startY);
-            points = this.currentPath.shape.points;
-        }
-
-        const theta = sweep / (segs * 2);
-        const theta2 = theta * 2;
-
-        const cTheta = Math.cos(theta);
-        const sTheta = Math.sin(theta);
-
-        const segMinus = segs - 1;
-
-        const remainder = (segMinus % 1) / segMinus;
-
-        for (let i = 0; i <= segMinus; ++i)
-        {
-            const real = i + (remainder * i);
-
-            const angle = ((theta) + startAngle + (theta2 * real));
-
-            const c = Math.cos(angle);
-            const s = -Math.sin(angle);
-
-            points.push(
-                (((cTheta * c) + (sTheta * s)) * radius) + cx,
-                (((cTheta * -s) + (sTheta * c)) * radius) + cy
-            );
-        }
-
-        this.dirty++;
-
-        return this;
-    }
-
-    /**
-     * Specifies a simple one-color fill that subsequent calls to other Graphics methods
-     * (such as lineTo() or drawCircle()) use when drawing.
-     *
-     * @param {number} [color=0] - the color of the fill
-     * @param {number} [alpha=1] - the alpha of the fill
-     * @return {PIXI.Graphics} This Graphics object. Good for chaining method calls
-     */
-    beginFill(color = 0, alpha = 1)
-    {
-        this.beginTextureFill(null, color, alpha);
-
-        return this;
-    }
-
-    beginTextureFill(texture, color = 0xFFFFFF, alpha = 1, textureMatrix)
-    {
-        if (alpha === 0)
-        {
-            this._fillStyle = null;
-
-            return this;
-        }
-
-        const style = new FillStyle();
-
-        style.color = color;
-        style.alpha = alpha;
-        style.texture = texture || Texture.WHITE;
-        style.matrix = textureMatrix;
-
-        this.filling = true;
-
-        if (this.currentPath)
-        {
-            if (this.currentPath.shape.points.length <= 2)
-            {
-                this.currentPath.fillStyle = style;
-            }
-        }
-
-        this._fillStyle = style;
-
-        return this;
-    }
-
-    /**
-     * Applies a fill to the lines and shapes that were added since the last call to the beginFill() method.
-     *
-     * @return {PIXI.Graphics} This Graphics object. Good for chaining method calls
-     */
-    endFill()
-    {
-        this._fillStyle = null;
-
-        return this;
-    }
-
-    /**
-     *
-     * @param {number} x - The X coord of the top-left of the rectangle
-     * @param {number} y - The Y coord of the top-left of the rectangle
-     * @param {number} width - The width of the rectangle
-     * @param {number} height - The height of the rectangle
-     * @return {PIXI.Graphics} This Graphics object. Good for chaining method calls
-     */
-    drawRect(x, y, width, height)
-    {
-        this.drawShape(new Rectangle(x, y, width, height), this._fillStyle, this._lineStyle, this.matrix);
-
-        return this;
-    }
-
-    /**
-     *
-     * @param {number} x - The X coord of the top-left of the rectangle
-     * @param {number} y - The Y coord of the top-left of the rectangle
-     * @param {number} width - The width of the rectangle
-     * @param {number} height - The height of the rectangle
-     * @param {number} radius - Radius of the rectangle corners
-     * @return {PIXI.Graphics} This Graphics object. Good for chaining method calls
-     */
-    drawRoundedRect(x, y, width, height, radius)
-    {
-        this.drawShape(new RoundedRectangle(x, y, width, height, radius), this._fillStyle, this._lineStyle, this.matrix);
-
-        return this;
-    }
-
-    /**
-     * Draws a circle.
-     *
-     * @param {number} x - The X coordinate of the center of the circle
-     * @param {number} y - The Y coordinate of the center of the circle
-     * @param {number} radius - The radius of the circle
-     * @return {PIXI.Graphics} This Graphics object. Good for chaining method calls
-     */
-    drawCircle(x, y, radius)
-    {
-        this.drawShape(new Circle(x, y, radius), this._fillStyle, this._lineStyle, this.matrix);
-
-        return this;
-    }
-
-    /**
-     * Draws an ellipse.
-     *
-     * @param {number} x - The X coordinate of the center of the ellipse
-     * @param {number} y - The Y coordinate of the center of the ellipse
-     * @param {number} width - The half width of the ellipse
-     * @param {number} height - The half height of the ellipse
-     * @return {PIXI.Graphics} This Graphics object. Good for chaining method calls
-     */
-    drawEllipse(x, y, width, height)
-    {
-        this.drawShape(new Ellipse(x, y, width, height), this._fillStyle, this._lineStyle, this.matrix);
-
-        return this;
-    }
-
-    /**
-     * Draws a polygon using the given path.
-     *
-     * @param {number[]|PIXI.Point[]|PIXI.Polygon} path - The path data used to construct the polygon.
-     * @return {PIXI.Graphics} This Graphics object. Good for chaining method calls
-     */
-    drawPolygon(path)
-    {
-        // prevents an argument assignment deopt
-        // see section 3.1: https://github.com/petkaantonov/bluebird/wiki/Optimization-killers#3-managing-arguments
-        let points = path;
-
-        let closed = true;
-
-        if (points instanceof Polygon)
-        {
-            closed = points.closed;
-            points = points.points;
-        }
-
-        if (!Array.isArray(points))
-        {
-            // prevents an argument leak deopt
-            // see section 3.2: https://github.com/petkaantonov/bluebird/wiki/Optimization-killers#3-managing-arguments
-            points = new Array(arguments.length);
-
-            for (let i = 0; i < points.length; ++i)
-            {
-                points[i] = arguments[i]; // eslint-disable-line prefer-rest-params
-            }
-        }
-
-        const shape = new Polygon(points);
-
-        shape.closed = closed;
-
-        this.drawShape(shape, this._fillStyle, this._lineStyle, this.matrix);
-
-        return this;
-    }
-
-    /**
-     * Draw a star shape with an arbitrary number of points.
-     *
-     * @param {number} x - Center X position of the star
-     * @param {number} y - Center Y position of the star
-     * @param {number} points - The number of points of the star, must be > 1
-     * @param {number} radius - The outer radius of the star
-     * @param {number} [innerRadius] - The inner radius between points, default half `radius`
-     * @param {number} [rotation=0] - The rotation of the star in radians, where 0 is vertical
-     * @return {PIXI.Graphics} This Graphics object. Good for chaining method calls
-     */
-    drawStar(x, y, points, radius, innerRadius, rotation = 0)
-    {
-        innerRadius = innerRadius || radius / 2;
-
-        const startAngle = (-1 * Math.PI / 2) + rotation;
-        const len = points * 2;
-        const delta = PI_2 / len;
-        const polygon = [];
-
-        for (let i = 0; i < len; i++)
-        {
-            const r = i % 2 ? innerRadius : radius;
-            const angle = (i * delta) + startAngle;
-
-            polygon.push(
-                x + (r * Math.cos(angle)),
-                y + (r * Math.sin(angle))
-            );
-        }
-
-        return this.drawPolygon(polygon);
-    }
-
-    /**
      * Clears the graphics that were drawn to this Graphics object, and resets fill and line style settings.
      *
      * @return {PIXI.Graphics} This Graphics object. Good for chaining method calls
      */
     clear()
     {
-        if (this.lineWidth || this.filling || this.graphicsData.length > 0)
+        if (this.graphicsData.length > 0)
         {
-            this._lineStyle = null;
-            this._fillStyle = null;
-
             this.boundsDirty = -1;
             this.dirty++;
             this.clearDirty++;
             this.graphicsData.length = 0;
-            this.graphicsDataHoles.length = 0;
-        }
 
-        this.currentPath = null;
+            this.drawCalls.length = 0;
+
+            this.points.length = 0;
+            this.colors.length = 0;
+            this.uvs.length = 0;
+
+            /**
+             * The indices of the vertices
+             * @member {number[]}
+             */
+            this.indices.length = 0;
+        }
 
         return this;
     }
@@ -706,6 +159,7 @@ export default class GraphicsGeometry extends Geometry
      */
     drawShape(shape, fillStyle, lineStyle, matrix)
     {
+        /*
         if (this.currentPath)
         {
             // check current path!
@@ -714,8 +168,9 @@ export default class GraphicsGeometry extends Geometry
                 this.graphicsData.pop();
             }
         }
+        */
 
-        this.currentPath = null;
+        //        this.currentPath = null;
 
         const data = new GraphicsData(
             shape,
@@ -724,68 +179,42 @@ export default class GraphicsGeometry extends Geometry
             matrix
         );
 
-        if (this.holeMode)
-        {
-            this.graphicsDataHoles.push(data);
-        }
-        else
-        {
-            this.graphicsData.push(data);
-        }
-
+        this.graphicsData.push(data);
+        /*
         if (data.type === SHAPES.POLY)
         {
             data.shape.closed = data.shape.closed || this.filling;
             this.currentPath = data;
         }
+*/
 
         this.dirty++;
-
-        return data;
-    }
-
-    /**
-     * Closes the current path.
-     *
-     * @return {PIXI.Graphics} Returns itself.
-     */
-    closePath()
-    {
-        // ok so close path assumes next one is a hole!
-        const currentPath = this.currentPath;
-
-        if (currentPath && currentPath.shape)
-        {
-            currentPath.shape.close();
-        }
 
         return this;
     }
 
     /**
-     * Begin adding holes to the last draw shape
-     * IMPORTANT: holes must be fully inside a shape to work
-     * Also weirdness ensues if holes overlap!
+     * Draws the given shape to this Graphics object. Can be any of Circle, Rectangle, Ellipse, Line or Polygon.
+     *
+     * @param {PIXI.Circle|PIXI.Ellipse|PIXI.Polygon|PIXI.Rectangle|PIXI.RoundedRectangle} shape - The shape object to draw.
+     * @return {PIXI.GraphicsData} The generated GraphicsData object.
      */
-    beginHole()
+    drawHole(shape, matrix)
     {
-        this.holeMode = true;
-    }
+        if (!this.graphicsData.length) return;
 
-    /**
-     * End adding holes to the last draw shape
-     */
-    endHole()
-    {
-        const currentPath = this.graphicsData[this.graphicsData.length - 1];
+        const data = new GraphicsData(
+            shape,
+            null,
+            null,
+            matrix
+        );
 
-        for (let i = 0; i < this.graphicsDataHoles.length; i++)
-        {
-            currentPath.holes[i] = this.graphicsDataHoles[i];
-        }
+        const lastShape = this.graphicsData[this.graphicsData.length - 1];
 
-        this.graphicsDataHoles.length = 0;
-        this.holeMode = false;
+        lastShape.holes.push(data);
+
+        this.dirty++;
     }
 
     /**
