+++ packages/graphics/src/shaders/primitive.vert
@@ -13,6 +13,8 @@ varying vec2 vTextureCoord;
 
 void main(void){
    gl_Position = vec4((projectionMatrix * translationMatrix * vec3(aVertexPosition, 1.0)).xy, 0.0, 1.0);
-   vColor = aColor * vec4(tint * alpha, alpha);
+   vColor = aColor;
+   vColor.rgb *= aColor.a;
+   vColor *= vec4(tint * alpha, alpha);
    vTextureCoord = aTextureCoord;
 }
